fix graph class to handle vertices by number
- vertices can now be addressed by their number using the getVertex method
- this actually slows down the entire class, but makes it possible to
	delete vertices from the graph without affecting the data structure
	or methods operating on this data

+++ src/Vertex.cpp
@@ -9,7 +9,7 @@
 #include "Vertex.hpp"
 
 // constructor
-Vertex::Vertex(const int number, const std::vector<int> &neighbors) : number((unsigned long) number), neighbors(neighbors) , color(0) {}
+Vertex::Vertex(const int number, const std::vector<int> &neighbors) : number( number), neighbors(neighbors) , color(0) {}
 
 // destructor
 Vertex::~Vertex() {}
